修改定时任务日志为定时任务事件日志

+++ server/crates/core/entity/src/log_api_operation.rs
@@ -35,7 +35,7 @@ pub struct Model {
     /// 用户代理
     pub user_agent: String,
     /// 耗时,毫秒
-    pub cost: f64,
+    pub cost: u64,
     /// 请求类型:REQ/RSP
     pub http_type: String,
     /// 备注
